allow disabling observatorium, pagerduty and deadmanssnitch

+++ controllers/reconcilers/promtail_installation/promtail_installation_reconciler.go
@@ -25,6 +25,12 @@ func NewReconciler(client client.Client, logger logr.Logger) reconcilers.Observa
 }
 
 func (r *Reconciler) Cleanup(ctx context.Context, cr *v1.Observability) (v1.ObservabilityStageStatus, error) {
+	// Without Observatorium there is no need to install Promtail, because we're not
+	// running on cluster Loki
+	if cr.ObservatoriumDisabled() || cr.ExternalSyncDisabled() {
+		return v1.ResultSuccess, nil
+	}
+
 	rolebinding := model.GetPromtailClusterRoleBinding(cr)
 	err := r.client.Delete(ctx, rolebinding)
 	if err != nil && !errors.IsNotFound(err) {
@@ -47,6 +53,12 @@ func (r *Reconciler) Cleanup(ctx context.Context, cr *v1.Observability) (v1.Obse
 }
 
 func (r *Reconciler) Reconcile(ctx context.Context, cr *v1.Observability, s *v1.ObservabilityStatus) (v1.ObservabilityStageStatus, error) {
+	// Without Observatorium there is no need to install Promtail, because we're not
+	// running on cluster Loki
+	if cr.ObservatoriumDisabled() || cr.ExternalSyncDisabled() {
+		return v1.ResultSuccess, nil
+	}
+
 	status, err := r.reconcilePromtailServiceAccount(ctx, cr)
 	if status != v1.ResultSuccess {
 		return status, err
